LayoutTests/svg/animations/animate-calcMode-spline-from-to-expected.txt

SVG 1.1 dynamic animation tests

Test calcMode spline with from-to animation. You should see a green 100x100 rect and only PASS messages

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 18.8
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 0
PASS rect.x.baseVal.value is 100
PASS rect.x.animVal.value is 100
PASS rect.x.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
